requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

streamlit
pillow
requests
thefuzz
python-Levenshtein
